+++ src/SecucardConnect/Product/Payment/Model/Basket.php
@@ -2,18 +2,17 @@
 
 namespace SecucardConnect\Product\Payment\Model;
 
-use SecucardConnect\Product\Common\Model\BaseModel;
-
 /**
  * Basket Data Model class
  *
  */
-class Basket extends BaseModel
+class Basket
 {
     const ITEM_TYPE_ARTICLE = 'article';
     const ITEM_TYPE_SHIPPING = 'shipping';
     const ITEM_TYPE_DONATION = 'donation';
     const ITEM_TYPE_STAKEHOLDER_PAYMENT = 'stakeholder_payment';
+    const ITEM_TYPE_SUB_TRANSACTION = 'sub_transaction';
 
     /**
      * @var int
@@ -70,4 +69,11 @@ class Basket extends BaseModel
      * @var string
      */
     public $apikey;
+
+    /**
+     * The basket of the sub-transaction
+     *
+     * @var Basket
+     */
+    public $sub_basket;
 }