+++ src/SecucardConnect/Product/Payment/Model/RedirectUrl.php
@@ -28,4 +28,11 @@ class RedirectUrl
      * @var string
      */
     public $iframe_url;
+    
+    /**
+     * Your endpoint to receive push notifications (when the status of the payment transaction will be changed).
+     *
+     * @var string
+     */
+    public $url_push;
 }