+++ src/SecucardConnect/Product/Payment/Model/SecupayPrepay.php
@@ -1,7 +1,4 @@
 <?php
-/**
- * Payment Secupayprepays Api Model class
- */
 
 namespace SecucardConnect\Product\Payment\Model;
 
@@ -20,34 +17,5 @@ class SecupayPrepay extends Transaction
      * @var TransferAccount
      */
     public $transfer_account;
-
-}
-
-class TransferAccount
-{
-    /**
-     * @var string
-     */
-    public $account_owner;
-
-    /**
-     * @var string
-     */
-    public $accountnumber;
-
-    /**
-     * @var string
-     */
-    public $iban;
-
-    /**
-     * @var string
-     */
-    public $bic;
-
-    /**
-     * @var string
-     */
-    public $bankcode;
 
 }