+++ src/SecucardConnect/Product/Payment/SecupayPayoutService.php
@@ -1,0 +1,15 @@
+<?php
+
+namespace SecucardConnect\Product\Payment;
+
+use SecucardConnect\Client\ProductService;
+
+/**
+ * Operations for the payment.secupaypayout resource.
+ * @package SecucardConnect\Product\Payment
+ */
+class SecupayPayoutService extends ProductService
+{
+
+}
+